requirements.txt
google-adk[fastapi]>=0.0.1
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
fastapi
uvicorn
python-dotenv
requests
wikipedia
pydantic>=2.0.0
